requirements.txt
streamlit
langchain
langchain_google_genai
python-docx
python-dotenv
pymupdf
pandas